gcp_functions/requirements.txt
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-bigquery==2.6.1
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
pandas==1.1.3
pyarrow==2.0.0
numpy==1.19.4